pyproject.toml
[tool.poetry]
name = "mygobot"
version = "0.1.0"
description = ""
authors = ["felix920506 <felix920506@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
py-cord = "^2.6.0"
pyyaml = "^6.0.2"
aiohttp = "^3.10.5"
aiohttp-client-cache = "^0.11.1"
aiofiles = "^24.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = "mygobot:main"